Use dnspython types for updates in all frontends
